+++ link.go
@@ -61,10 +61,6 @@ type link struct {
 	// can independently set this value. The sender endpoint sets this to the last known value seen from the receiver.
 	linkCredit uint32
 
-	// The number of messages awaiting credit at the link sender endpoint. Only the sender can independently
-	// set this value. The receiver sets this to the last known value seen from the sender.
-	availableCredit uint32
-
 	senderSettleMode   *SenderSettleMode
 	receiverSettleMode *ReceiverSettleMode
 	maxMessageSize     uint64