+++ sender.go
@@ -284,11 +284,11 @@ func (s *Sender) mux() {
 Loop:
 	for {
 		var outgoingTransfers chan frames.PerformTransfer
-		if s.l.availableCredit > 0 {
-			debug.Log(1, "TX (Sender) (enable): target: %q, available credit: %d, deliveryCount: %d", s.l.target.Address, s.l.availableCredit, s.l.deliveryCount)
+		if s.l.linkCredit > 0 {
+			debug.Log(1, "TX (Sender) (enable): target: %q, link credit: %d, deliveryCount: %d", s.l.target.Address, s.l.linkCredit, s.l.deliveryCount)
 			outgoingTransfers = s.transfers
 		} else {
-			debug.Log(1, "TX (Sender) (pause): target: %q, available credit: %d, deliveryCount: %d", s.l.target.Address, s.l.availableCredit, s.l.deliveryCount)
+			debug.Log(1, "TX (Sender) (pause): target: %q, link credit: %d, deliveryCount: %d", s.l.target.Address, s.l.linkCredit, s.l.deliveryCount)
 		}
 
 		select {
@@ -310,9 +310,9 @@ Loop:
 				// decrement link-credit after entire message transferred
 				if !tr.More {
 					s.l.deliveryCount++
-					s.l.availableCredit--
+					s.l.linkCredit--
 					// we are the sender and we keep track of the peer's link credit
-					debug.Log(3, "TX (Sender): link: %s, available credit: %d", s.l.key.name, s.l.availableCredit)
+					debug.Log(3, "TX (Sender): link: %s, link credit: %d", s.l.key.name, s.l.linkCredit)
 				}
 				continue Loop
 			case <-s.l.close:
@@ -339,6 +339,8 @@ func (s *Sender) muxHandleFrame(fr frames.FrameBody) error {
 	switch fr := fr.(type) {
 	// flow control frame
 	case *frames.PerformFlow:
+		// the sender's link-credit variable MUST be set according to this formula when flow information is given by the receiver:
+		// link-credit(snd) := delivery-count(rcv) + link-credit(rcv) - delivery-count(snd)
 		linkCredit := *fr.LinkCredit - s.l.deliveryCount
 		if fr.DeliveryCount != nil {
 			// DeliveryCount can be nil if the receiver hasn't processed
@@ -346,8 +348,8 @@ func (s *Sender) muxHandleFrame(fr frames.FrameBody) error {
 			// what ActiveMQ does.
 			linkCredit += *fr.DeliveryCount
 		}
-		// TODO: clean up as part of flow control fixes
-		s.l.availableCredit = linkCredit
+
+		s.l.linkCredit = linkCredit
 
 		if !fr.Echo {
 			return nil
@@ -359,7 +361,6 @@ func (s *Sender) muxHandleFrame(fr frames.FrameBody) error {
 		)
 
 		// send flow
-		// TODO: missing Available
 		resp := &frames.PerformFlow{
 			Handle:        &s.l.handle,
 			DeliveryCount: &deliveryCount,